Move Minecraft skin control out of builder into _common/skins
The engine and ui layers should be independent of any generator-specific
details, Minecraft included. The skin picker, its fetch/conversion
logic, and the MinecraftSkinOption types were the last Minecraft-aware
code left inside src/builder — nothing generic in the framework
(engine.ts, model.ts, generator.ts, generatorUI.tsx,
generatorRenderer.tsx, renderContextAdapter.ts) ever depended on them,
so the move is a relocation, not a redesign.

minecraftSkin.ts, minecraftSkinConverter.ts, minecraftSkinInputValue.ts,
and minecraftSkinControl.tsx move to src/generators/_common/skins/,
following the same pattern _common/tintSelector already uses for
reaching generic ui/ primitives (Select, Button, icons) via deep
@genroot/builder/ui/* imports. The MinecraftSkinOption/Preset/Texture
types move into the existing options.ts rather than a new file.

This also removes the circular import modelControls.ts <-> the old
minecraftSkinInputValue.ts had: the Option types no longer live
alongside RegionControl, so modelControls.ts drops its
minecraftSkinInputValue.ts re-export entirely and has zero
Minecraft-related references. A second, smaller cycle the move would
otherwise have introduced inside _common/skins/ (index.ts ->
minecraftSkinInputValue.ts -> options.ts -> index.ts) is avoided by
inlining the two-value "Wide" | "Slim" union in index.ts's getSkinUrl
instead of importing MinecraftModelType for it.

skinControl.ts keeps its role as the one stable import path the 20
Minecraft generators already use, now re-exporting local siblings
instead of reaching into builder.

Verified with types:check, lint, check:imports, the full vitest suite,
and the full Playwright generator suite (run against this worktree's
own dev server on an isolated port, since another concurrent session
was already occupying the default port 3001).
